@@ -7,27 +7,28 @@
     },
     "description": "Converts your TSLint configuration to the closest possible ESLint equivalent.",
     "dependencies": {
-        "chalk": "2.4.2",
-        "commander": "3.0.2",
-        "tslint": "5.20.0",
-        "typescript": "3.6.3"
+        "chalk": "3.0.0",
+        "commander": "4.0.1",
+        "strip-json-comments": "^3.0.1",
+        "tslint": "5.20.1",
+        "typescript": "3.7.2"
     },
     "devDependencies": {
-        "@babel/core": "7.6.2",
-        "@babel/preset-env": "7.6.2",
-        "@babel/preset-typescript": "7.6.0",
-        "@types/jest": "24.0.18",
-        "@types/node": "12.7.8",
-        "@typescript-eslint/eslint-plugin": "2.3.2",
-        "@typescript-eslint/parser": "2.3.2",
+        "@babel/core": "7.7.2",
+        "@babel/preset-env": "7.7.1",
+        "@babel/preset-typescript": "7.7.2",
+        "@types/jest": "24.0.23",
+        "@types/node": "12.12.7",
+        "@typescript-eslint/eslint-plugin": "2.7.0",
+        "@typescript-eslint/parser": "2.7.0",
         "babel-jest": "24.9.0",
-        "eslint": "6.5.1",
-        "eslint-config-prettier": "6.4.0",
-        "husky": "3.0.7",
+        "eslint": "6.6.0",
+        "eslint-config-prettier": "6.5.0",
+        "husky": "3.0.9",
         "jest": "24.9.0",
-        "lint-staged": "9.4.0",
-        "prettier": "1.18.2",
-        "strip-ansi": "5.2.0"
+        "lint-staged": "9.4.3",
+        "prettier": "1.19.1",
+        "strip-ansi": "6.0.0"
     },
     "homepage": "https://github.com/typescript-eslint/tslint-to-eslint-config#readme",
     "husky": {
@@ -60,5 +61,5 @@
         "test": "jest",
         "tsc": "tsc"
     },
-    "version": "0.2.7"
+    "version": "0.3.1"
 }
