@@ -1,4 +1,7 @@
 import { EOL } from "os";
 
-export const formatJsOutput = (configuration: unknown) =>
-    `module.exports = ${JSON.stringify(configuration, undefined, 4)};${EOL}`;
+import { withKeysSorted } from "./withKeysSorted";
+
+export const formatJsOutput = (configuration: any) => {
+    return `module.exports = ${JSON.stringify(withKeysSorted(configuration), undefined, 4)};${EOL}`;
+};
