@@ -3,6 +3,7 @@ import { RuleConversionResults } from "../../rules/convertRules";
 import { removeExtendsDuplicatedRules } from "./removeExtendsDuplicatedRules";
 import { retrieveExtendsValues } from "./retrieveExtendsValues";
 import { ESLintConfiguration } from "../../input/findESLintConfiguration";
+import { OriginalConfigurations } from "../../input/findOriginalConfigurations";
 
 export type SimplifyPackageRulesDependencies = {
     removeExtendsDuplicatedRules: typeof removeExtendsDuplicatedRules;
@@ -13,15 +14,15 @@ export type SimplifiedRuleConversionResults = Pick<RuleConversionResults, "conve
 
 export const simplifyPackageRules = async (
     dependencies: SimplifyPackageRulesDependencies,
-    eslint: Partial<ESLintConfiguration> | undefined,
+    eslint: Pick<OriginalConfigurations<ESLintConfiguration>, "full"> | undefined,
     ruleConversionResults: SimplifiedRuleConversionResults,
 ): Promise<SimplifiedRuleConversionResults> => {
-    if (eslint === undefined || eslint.extends === undefined || eslint.extends.length === 0) {
+    if (eslint === undefined || eslint.full.extends.length === 0) {
         return ruleConversionResults;
     }
 
     const { configurationErrors, importedExtensions } = await dependencies.retrieveExtendsValues(
-        eslint.extends,
+        eslint.full.extends,
     );
 
     const converted = dependencies.removeExtendsDuplicatedRules(
