@@ -1,12 +1,15 @@
 import { createEmptyConversionResults } from "../conversion/conversionResults.stubs";
 import { writeConversionResults } from "./writeConversionResults";
-import { OriginalConfigurations } from "../input/findOriginalConfigurations";
+import { AllOriginalConfigurations } from "../input/findOriginalConfigurations";
 import { formatJsonOutput } from "./formatting/formatters/formatJsonOutput";
 
-const createStubOriginalConfigurations = (overrides: Partial<OriginalConfigurations> = {}) => ({
+const createStubOriginalConfigurations = (overrides: Partial<AllOriginalConfigurations> = {}) => ({
     tslint: {
-        rulesDirectory: [],
-        rules: {},
+        full: {
+            rulesDirectory: [],
+            rules: {},
+        },
+        raw: {},
     },
     ...overrides,
 });
@@ -14,9 +17,7 @@ const createStubOriginalConfigurations = (overrides: Partial<OriginalConfigurati
 describe("writeConversionResults", () => {
     it("excludes the tslint plugin when there are no missing rules", async () => {
         // Arrange
-        const conversionResults = createEmptyConversionResults({
-            converted: new Map(),
-        });
+        const conversionResults = createEmptyConversionResults();
         const fileSystem = { writeFile: jest.fn().mockReturnValue(Promise.resolve()) };
 
         // Act
@@ -50,7 +51,6 @@ describe("writeConversionResults", () => {
     it("includes typescript-eslint plugin settings when there are missing rules", async () => {
         // Arrange
         const conversionResults = createEmptyConversionResults({
-            converted: new Map(),
             missing: [
                 {
                     ruleArguments: [],
@@ -100,16 +100,17 @@ describe("writeConversionResults", () => {
 
     it("includes the original eslint configuration when it exists", async () => {
         // Arrange
-        const conversionResults = createEmptyConversionResults({
-            converted: new Map(),
-        });
+        const conversionResults = createEmptyConversionResults();
         const eslint = {
-            env: {},
-            extends: [],
-            globals: {
-                Promise: true,
+            full: {
+                env: {},
+                extends: [],
+                globals: {
+                    Promise: true,
+                },
+                rules: {},
             },
-            rules: {},
+            raw: {},
         };
         const originalConfigurations = createStubOriginalConfigurations({
             eslint,
@@ -128,11 +129,64 @@ describe("writeConversionResults", () => {
         expect(fileSystem.writeFile).toHaveBeenLastCalledWith(
             ".eslintrc.json",
             formatJsonOutput({
-                ...eslint,
                 env: {
                     browser: true,
                     es6: true,
                     node: true,
+                },
+                extends: [],
+                rules: {},
+                parser: "@typescript-eslint/parser",
+                parserOptions: {
+                    project: "tsconfig.json",
+                    sourceType: "module",
+                },
+                plugins: ["@typescript-eslint"],
+            }),
+        );
+    });
+
+    it("includes raw globals when they exist", async () => {
+        // Arrange
+        const conversionResults = createEmptyConversionResults();
+        const eslint = {
+            full: {
+                env: {},
+                extends: [],
+                rules: {},
+            },
+            raw: {
+                globals: {
+                    Promise: true,
+                },
+            },
+        };
+        const originalConfigurations = createStubOriginalConfigurations({
+            eslint,
+        });
+        const fileSystem = { writeFile: jest.fn().mockReturnValue(Promise.resolve()) };
+
+        // Act
+        await writeConversionResults(
+            { fileSystem },
+            ".eslintrc.json",
+            conversionResults,
+            originalConfigurations,
+        );
+
+        // Assert
+        expect(fileSystem.writeFile).toHaveBeenLastCalledWith(
+            ".eslintrc.json",
+            formatJsonOutput({
+                env: {
+                    browser: true,
+                    es6: true,
+                    node: true,
+                },
+                extends: [],
+                rules: {},
+                globals: {
+                    Promise: true,
                 },
                 parser: "@typescript-eslint/parser",
                 parserOptions: {
@@ -140,7 +194,6 @@ describe("writeConversionResults", () => {
                     sourceType: "module",
                 },
                 plugins: ["@typescript-eslint"],
-                rules: {},
             }),
         );
     });
