@@ -1,6 +1,15 @@
-import { findESLintConfiguration } from "./findESLintConfiguration";
+import {
+    findESLintConfiguration,
+    FindESLintConfigurationDependencies,
+} from "./findESLintConfiguration";
 import { createStubExec, createStubThrowingExec } from "../adapters/exec.stubs";
 import { TSLintToESLintSettings } from "../types";
+
+const createStubDependencies = (overrides: Partial<FindESLintConfigurationDependencies> = {}) => ({
+    exec: createStubExec({ stdout: "{}" }),
+    importer: async () => ({}),
+    ...overrides,
+});
 
 const createStubRawSettings = (overrides: Partial<TSLintToESLintSettings> = {}) => ({
     config: "./eslintrc.js",
@@ -9,10 +18,32 @@ const createStubRawSettings = (overrides: Partial<TSLintToESLintSettings> = {}) 
 });
 
 describe("findESLintConfiguration", () => {
-    it("returns an error when one occurs", async () => {
+    it("returns an error when exec returns one", async () => {
         // Arrange
         const message = "error";
-        const dependencies = { exec: createStubThrowingExec({ stderr: message }) };
+        const dependencies = createStubDependencies({
+            exec: createStubThrowingExec({ stderr: message }),
+        });
+
+        // Act
+        const result = await findESLintConfiguration(dependencies, createStubRawSettings());
+
+        // Assert
+        expect(result).toEqual(
+            expect.objectContaining({
+                message,
+            }),
+        );
+    });
+
+    it("returns an error when importer returns one", async () => {
+        // Arrange
+        const message = "error";
+        const dependencies = createStubDependencies({
+            importer: async () => {
+                throw new Error(message);
+            },
+        });
 
         // Act
         const result = await findESLintConfiguration(dependencies, createStubRawSettings());
@@ -27,7 +58,7 @@ describe("findESLintConfiguration", () => {
 
     it("defaults the configuration file when one isn't provided", async () => {
         // Arrange
-        const dependencies = { exec: createStubExec() };
+        const dependencies = createStubDependencies({ exec: createStubExec() });
 
         // Act
         await findESLintConfiguration(dependencies, createStubRawSettings());
@@ -38,7 +69,7 @@ describe("findESLintConfiguration", () => {
 
     it("includes a configuration file in the ESLint command when one is provided", async () => {
         // Arrange
-        const dependencies = { exec: createStubExec() };
+        const dependencies = createStubDependencies({ exec: createStubExec() });
         const config = createStubRawSettings({
             eslint: "./custom/eslintrc.js",
         });
@@ -54,7 +85,7 @@ describe("findESLintConfiguration", () => {
 
     it("applies ESLint defaults when none are provided", async () => {
         // Arrange
-        const dependencies = { exec: createStubExec({ stdout: "{}" }) };
+        const dependencies = createStubDependencies({ exec: createStubExec({ stdout: "{}" }) });
         const config = createStubRawSettings({
             eslint: "./custom/eslintrc.js",
         });
@@ -64,9 +95,77 @@ describe("findESLintConfiguration", () => {
 
         // Assert
         expect(result).toEqual({
-            env: {},
-            extends: [],
-            rules: {},
+            full: {
+                env: {},
+                extends: [],
+                rules: {},
+            },
+            raw: {
+                extends: [],
+            },
+        });
+    });
+
+    it("doesn't apply raw extends on top of reported when they don't exist", async () => {
+        // Arrange
+        const reportedExtends = ["reported"];
+        const dependencies = createStubDependencies({
+            exec: createStubExec({
+                stdout: JSON.stringify({
+                    extends: reportedExtends,
+                }),
+            }),
+            importer: async () => ({}),
+        });
+        const config = createStubRawSettings({
+            eslint: "./custom/eslintrc.js",
+        });
+
+        // Act
+        const result = await findESLintConfiguration(dependencies, config);
+
+        // Assert
+        expect(result).toEqual({
+            full: {
+                env: {},
+                extends: ["reported"],
+                rules: {},
+            },
+            raw: {
+                extends: [],
+            },
+        });
+    });
+
+    it("applies raw extends on top of reported when they exist", async () => {
+        // Arrange
+        const raw = {
+            extends: ["raw", "duplicated"],
+        };
+        const reportedExtends = ["reported", "duplicated"];
+        const dependencies = createStubDependencies({
+            exec: createStubExec({
+                stdout: JSON.stringify({
+                    extends: reportedExtends,
+                }),
+            }),
+            importer: async () => raw,
+        });
+        const config = createStubRawSettings({
+            eslint: "./custom/eslintrc.js",
+        });
+
+        // Act
+        const result = await findESLintConfiguration(dependencies, config);
+
+        // Assert
+        expect(result).toEqual({
+            full: {
+                env: {},
+                extends: ["raw", "duplicated", "reported"],
+                rules: {},
+            },
+            raw,
         });
     });
 });
