@@ -17,17 +17,32 @@ export type FindOriginalConfigurationsDependencies = {
     mergeLintConfigurations: typeof mergeLintConfigurations;
 };
 
-export type OriginalConfigurations = {
-    eslint?: ESLintConfiguration;
+/**
+ * Both found configurations for a particular linter.
+ */
+export type OriginalConfigurations<Configuration> = {
+    /**
+     * Settings reported by the linter's native --print-config equivalent.
+     */
+    full: Configuration;
+
+    /**
+     * Raw import results from `import`ing the configuration file.
+     */
+    raw: Partial<Configuration>;
+};
+
+export type AllOriginalConfigurations = {
+    eslint?: OriginalConfigurations<ESLintConfiguration>;
     packages?: PackagesConfiguration;
-    tslint: TSLintConfiguration;
+    tslint: OriginalConfigurations<TSLintConfiguration>;
     typescript?: TypeScriptConfiguration;
 };
 
 export const findOriginalConfigurations = async (
     dependencies: FindOriginalConfigurationsDependencies,
     rawSettings: TSLintToESLintSettings,
-): Promise<ResultWithDataStatus<OriginalConfigurations>> => {
+): Promise<ResultWithDataStatus<AllOriginalConfigurations>> => {
     // Simultaneously search for all required configuration types
     const [eslint, packages, tslint, typescript] = await Promise.all([
         dependencies.findESLintConfiguration(rawSettings),
