@@ -4,11 +4,11 @@ export type DeepPartial<T> = {
     [P in keyof T]: T[P] extends {} ? DeepPartial<T[P]> : T[P];
 };
 
-export type FindConfigurationDependencies = {
+export type FindReportedConfigurationDependencies = {
     exec: Exec;
 };
 
-export const findConfiguration = async <Configuration>(
+export const findReportedConfiguration = async <Configuration>(
     exec: Exec,
     command: string,
     config: string,
