@@ -1,44 +1,61 @@
+import { findRawConfiguration } from "./findRawConfiguration";
+import { findReportedConfiguration } from "./findReportedConfiguration";
 import { Exec } from "../adapters/exec";
-import { findConfiguration } from "./findConfiguration";
+import { SansDependencies } from "../binding";
+import { uniqueFromSources } from "../utils";
+import { importer } from "./importer";
 
 export type TSLintConfiguration = {
-    rulesDirectory: string[];
+    extends?: string[];
+    rulesDirectory?: string[];
     rules: TSLintConfigurationRules;
 };
 
-export type TSLintConfigurationRules = {
-    [i: string]: any;
-};
-
-const defaultTSLintConfiguration = {
-    rulesDirectory: [],
-    rules: {},
-};
+export type TSLintConfigurationRules = Record<string, any>;
 
 export type FindTSLintConfigurationDependencies = {
     exec: Exec;
+    importer: SansDependencies<typeof importer>;
 };
 
 export const findTSLintConfiguration = async (
     dependencies: FindTSLintConfigurationDependencies,
     config: string | undefined,
-): Promise<TSLintConfiguration | Error> => {
-    const rawConfiguration = await findConfiguration<TSLintConfiguration>(
-        dependencies.exec,
-        "tslint --print-config",
-        config || "./tslint.json",
-    );
+) => {
+    const filePath = config || "./tslint.json";
+    const [rawConfiguration, reportedConfiguration] = await Promise.all([
+        findRawConfiguration<Partial<TSLintConfiguration>>(dependencies.importer, filePath),
+        findReportedConfiguration<TSLintConfiguration>(
+            dependencies.exec,
+            "tslint --print-config",
+            config || "./tslint.json",
+        ),
+    ]);
 
-    if (rawConfiguration instanceof Error) {
-        if (rawConfiguration.message.includes("unknown option `--print-config")) {
+    if (reportedConfiguration instanceof Error) {
+        if (reportedConfiguration.message.includes("unknown option `--print-config")) {
             return new Error("TSLint v5.18 required. Please update your version.");
         }
 
+        return reportedConfiguration;
+    }
+
+    if (rawConfiguration instanceof Error) {
         return rawConfiguration;
     }
 
+    const extensions = uniqueFromSources(rawConfiguration.extends, reportedConfiguration.extends);
+
+    const rules = {
+        ...rawConfiguration.rules,
+        ...reportedConfiguration.rules,
+    };
+
     return {
-        ...defaultTSLintConfiguration,
-        ...rawConfiguration,
+        full: {
+            ...(extensions.length !== 0 && { extends: extensions }),
+            rules,
+        },
+        raw: rawConfiguration,
     };
 };
