@@ -1,24 +1,28 @@
-import { TSLintConfiguration } from "./findTSLintConfiguration";
 import { ESLintConfiguration } from "./findESLintConfiguration";
+import { OriginalConfigurations } from "./findOriginalConfigurations";
+import { TSLintConfiguration } from "./findTSLintConfiguration";
 
 export const mergeLintConfigurations = (
-    eslint: ESLintConfiguration | Error,
-    tslint: TSLintConfiguration,
-) => {
+    eslint: OriginalConfigurations<ESLintConfiguration> | Error,
+    tslint: OriginalConfigurations<TSLintConfiguration>,
+): OriginalConfigurations<TSLintConfiguration> => {
     if (eslint instanceof Error) {
         return tslint;
     }
 
-    const mappedConfig = eslint.rules["@typescript-eslint/tslint/config"];
+    const mappedConfig = eslint.full.rules["@typescript-eslint/tslint/config"];
     if (!(mappedConfig instanceof Array) || mappedConfig[0] === "off") {
         return tslint;
     }
 
     return {
         ...tslint,
-        rules: {
-            ...tslint.rules,
-            ...mappedConfig[1].rules,
+        full: {
+            ...tslint.full,
+            rules: {
+                ...tslint.full.rules,
+                ...mappedConfig[1].rules,
+            },
         },
     };
 };
