@@ -4,7 +4,7 @@ export const convertNoAngleBracketTypeAssertion: RuleConverter = () => {
     return {
         rules: [
             {
-                ruleName: "@typescript-eslint/no-angle-bracket-type-assertion",
+                ruleName: "@typescript-eslint/consistent-type-assertions",
             },
         ],
     };
