@@ -4,7 +4,7 @@ export const convertNoReference: RuleConverter = () => {
     return {
         rules: [
             {
-                ruleName: "@typescript-eslint/no-triple-slash-reference",
+                ruleName: "@typescript-eslint/triple-slash-reference",
             },
         ],
     };
