@@ -11,9 +11,9 @@ export const convertNoUnusedExpression: RuleConverter = tslintRule => {
     };
 };
 
-const collectNoticesAndArguments = (tsLintRuleArguments: any[]) => {
-    const noAllowNewNotice = `The TSLint optional config "allow-new" is the default ESLint behavior and will no longer be ignored.`;
+const noAllowNewNotice = `The TSLint optional config "allow-new" is the default ESLint behavior and will no longer be ignored.`;
 
+const collectNoticesAndArguments = (tsLintRuleArguments: any[]) => {
     if (tsLintRuleArguments.length === 0) {
         return {
             notices: [noAllowNewNotice],
@@ -36,8 +36,8 @@ const collectNoticesAndArguments = (tsLintRuleArguments: any[]) => {
     }
 
     return {
-        ...(notices.length > 0 && { notices }),
-        ...(ruleArguments.length > 0 && {
+        ...(notices.length !== 0 && { notices }),
+        ...(ruleArguments.length !== 0 && {
             ruleArguments: [
                 ruleArguments.reduce((value, current) => Object.assign(value, current), {}),
             ],
