@@ -4,10 +4,9 @@ export const convertNoVoidExpression: RuleConverter = tslintRule => {
     return {
         rules: [
             {
-                ...(tslintRule.ruleArguments.length > 0 &&
-                    tslintRule.ruleArguments.includes("ignore-arrow-function-shorthand") && {
-                        notices: ["ESLint does not support ignoring arrow function shorthand."],
-                    }),
+                ...(tslintRule.ruleArguments.includes("ignore-arrow-function-shorthand") && {
+                    notices: ["ESLint does not support ignoring arrow function shorthand."],
+                }),
                 ruleName: "no-void",
             },
         ],
