@@ -4,7 +4,9 @@ export const convertObjectLiteralKeyQuotes: RuleConverter = tslintRule => {
     return {
         rules: [
             {
-                ruleArguments: tslintRule.ruleArguments,
+                ...(tslintRule.ruleArguments.length !== 0 && {
+                    ruleArguments: tslintRule.ruleArguments,
+                }),
                 ruleName: "quote-props",
             },
         ],
