@@ -4,10 +4,9 @@ export const convertPreferReadonly: RuleConverter = tslintRule => {
     return {
         rules: [
             {
-                ...(tslintRule.ruleArguments.length !== 0 &&
-                    tslintRule.ruleArguments[0] === "only-inline-lambdas" && {
-                        ruleArguments: [{ onlyInlineLambdas: true }],
-                    }),
+                ...(tslintRule.ruleArguments.includes("only-inline-lambdas") && {
+                    ruleArguments: [{ onlyInlineLambdas: true }],
+                }),
                 ruleName: "prefer-readonly",
             },
         ],
