@@ -1,21 +1,22 @@
 import { RuleConverter } from "../converter";
 
 type QuotemarkRule = string | { avoidEscape: true };
+
 export const convertQuotemark: RuleConverter = tslintRule => {
     const notices: string[] = [];
     const ruleArguments: QuotemarkRule[] = [];
 
-    ["jsx-single", "jsx-double", "avoid-template"].forEach(option => {
+    for (const option of ["jsx-single", "jsx-double", "avoid-template"]) {
         if (tslintRule.ruleArguments.includes(option)) {
             notices.push(`Option "${option}" is not supported by ESLint.`);
         }
-    });
+    }
 
-    ["single", "double", "backtick"].forEach(option => {
+    for (const option of ["single", "double", "backtick"]) {
         if (tslintRule.ruleArguments.includes(option)) {
             ruleArguments.push(option);
         }
-    });
+    }
 
     if (tslintRule.ruleArguments.includes("avoid-escape")) {
         ruleArguments.push({ avoidEscape: true });
@@ -24,8 +25,8 @@ export const convertQuotemark: RuleConverter = tslintRule => {
     return {
         rules: [
             {
-                notices,
-                ruleArguments,
+                ...(notices.length !== 0 && { notices }),
+                ...(ruleArguments.length !== 0 && { ruleArguments }),
                 ruleName: "@typescript-eslint/quotes",
             },
         ],
